requirements.txt
azure-functions
azure-storage-blob
requests
azure-identity
azure-keyvault-secrets
